refactor(dynamic-agents): rename encoders/ to stream_encoders/ and absorb langgraph helpers
- encoders/ -> stream_encoders/ for clarity
- langgraph_stream_helpers.py -> stream_encoders/langgraph_helpers.py
- Update all import sites and test mock patch paths
