+++ ai_platform_engineering/dynamic_agents/src/dynamic_agents/services/stream_encoders/agui_sse.py
@@ -26,8 +26,8 @@ import time
 from typing import Any
 from uuid import uuid4
 
-from dynamic_agents.services.encoders import StreamEncoder
-from dynamic_agents.services.langgraph_stream_helpers import LangGraphStreamHelper
+from dynamic_agents.services.stream_encoders import StreamEncoder
+from dynamic_agents.services.stream_encoders.langgraph_helpers import LangGraphStreamHelper
 
 logger = logging.getLogger(__name__)
